@@ -12,11 +12,11 @@ import javax.xml.ws.WebServiceFeature;
 
 /**
  * This class was generated by the JAX-WS RI.
- * JAX-WS RI 2.3.2
+ * JAX-WS RI 2.2.9-b130926.1035
  * Generated source version: 2.2
  * 
  */
-@WebServiceClient(name = "CountryServiceImplService", targetNamespace = "http://server.ws.soap.baeldung.com/", wsdlLocation = "file:src/main/resources/country.wsdl")
+@WebServiceClient(name = "CountryServiceImplService", targetNamespace = "http://server.ws.soap.baeldung.com/", wsdlLocation = "http://localhost:8888/ws/country?wsdl")
 public class CountryServiceImplService extends Service {
 
     private final static URL COUNTRYSERVICEIMPLSERVICE_WSDL_LOCATION;
@@ -27,7 +27,7 @@ public class CountryServiceImplService extends Service {
         URL url = null;
         WebServiceException e = null;
         try {
-            url = new URL("file:src/main/resources/country.wsdl");
+            url = new URL("http://localhost:8888/ws/country?wsdl");
         } catch (MalformedURLException ex) {
             e = new WebServiceException(ex);
         }
