@@ -1,12 +1,13 @@
 package com.baeldung.spring.data.dynamodb.repositories;
 
-import com.baeldung.spring.data.dynamodb.model.ProductInfo;
+import java.util.Optional;
+
 import org.socialsignin.spring.data.dynamodb.repository.EnableScan;
 import org.springframework.data.repository.CrudRepository;
 
-import java.util.List;
+import com.baeldung.spring.data.dynamodb.model.ProductInfo;
 
 @EnableScan
 public interface ProductInfoRepository extends CrudRepository<ProductInfo, String> {
-    List<ProductInfo> findById(String id);
+    Optional<ProductInfo> findById(String id);
 }
